Sweeping changes to controllers to remove static fori loops. INtroduced new queries on database layer that controllers can now access for direct information.

+++ src/main/java/group03/project/repositories/ObjectiveRepoJPA.java
@@ -4,5 +4,5 @@ import group03.project.domain.Objective;
 import group03.project.services.required.ObjectiveRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObjectiveRepoJPA extends JpaRepository<Objective, Long>, ObjectiveRepository {
+public interface ObjectiveRepoJPA extends JpaRepository<Objective, Integer>, ObjectiveRepository {
 }
